+++ test/async-hooks/test-tcpwrap.js
@@ -6,7 +6,7 @@ if (!common.hasIPv6)
   common.skip('IPv6 support required');
 
 const assert = require('assert');
-const tick = require('./tick');
+const tick = require('../common/tick');
 const initHooks = require('./init-hooks');
 const { checkInvocations } = require('./hook-checks');
 const net = require('net');
