+++ test/async-hooks/test-timers.setTimeout.js
@@ -2,7 +2,7 @@
 
 const common = require('../common');
 const assert = require('assert');
-const tick = require('./tick');
+const tick = require('../common/tick');
 const initHooks = require('./init-hooks');
 const { checkInvocations } = require('./hook-checks');
 const TIMEOUT = common.platformTimeout(100);
