+++ test/async-hooks/test-ttywrap.writestream.js
@@ -4,7 +4,7 @@ const common = require('../common');
 const assert = require('assert');
 
 // general hook test setup
-const tick = require('./tick');
+const tick = require('../common/tick');
 const initHooks = require('./init-hooks');
 const { checkInvocations } = require('./hook-checks');
